# docs/openapi.yaml
openapi: 3.0.3
info:
  title: KORION Chong API
  version: 0.1.0
paths:
...
  /api/leader/dashboard:
    get:
      summary: Country leader dashboard summary
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - name: period
          in: query
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
          example: '2026-06'
        - name: countryScope
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 2
          example: KR
      responses:
        '200':
          description: Dashboard summary scoped to the authenticated country leader.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderDashboardResponse'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/partners:
    get:
      summary: Approved sales partners under a country leader
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - name: countryScope
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 2
          example: KR
        - name: keyword
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum:
              - SALES_PARTNER_APPROVED
        - name: region
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Approved sales partner page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderPartnerResponse'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/partner-applications:
    get:
      summary: Partner signup applications for leader request page
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/LeaderPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/partner-requests:
    get:
      summary: Alias for partner signup applications
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/LeaderPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/merchant-applications:
    get:
      summary: Merchant signup applications for leader request page
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/LeaderPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/merchant-requests:
    get:
      summary: Alias for merchant signup applications
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/LeaderPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/partner-sales:
    get:
      summary: Partner sales summary page data
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
        - $ref: '#/components/parameters/PeriodQuery'
      responses:
        '200':
          $ref: '#/components/responses/LeaderPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/merchants:
    get:
      summary: Merchants under a country leader
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/LeaderPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/merchant-sales:
    get:
      summary: Merchant sales summary page data
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
        - $ref: '#/components/parameters/PeriodQuery'
      responses:
        '200':
          $ref: '#/components/responses/LeaderPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/transactions:
    get:
      summary: Transaction log page data
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
        - $ref: '#/components/parameters/PeriodQuery'
        - name: variant
          in: query
          schema:
            type: string
            enum: [all, offline, failed]
            default: all
      responses:
        '200':
          $ref: '#/components/responses/LeaderPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/settlements/request-summary:
    get:
      summary: Settlement request summary page data
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
        - $ref: '#/components/parameters/PeriodQuery'
      responses:
        '200':
          $ref: '#/components/responses/LeaderPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/settlements:
    get:
      summary: Settlement history page data
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/LeaderPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/settlements/detail:
    get:
      summary: Settlement detail page data
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/LeaderPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/settlement-requests:
    post:
      summary: Create a country leader settlement request
      description: Recalculates available commission entries server-side and creates a REQUESTED distributor_settlement_requests row for the authenticated leader.
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettlementCreateRequest'
      responses:
        '200':
          description: Settlement request created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementActionResponse'
  /api/leader/settlement-requests/{settlementRequestId}/approve:
    post:
      summary: Approve a subordinate settlement request
      description: Moves a manageable REQUESTED settlement request to APPROVED and marks linked commission entries APPROVED or HELD.
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/SettlementRequestIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettlementActionRequest'
      responses:
        '200':
          description: Settlement request approved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementActionResponse'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/settlement-requests/{settlementRequestId}/reject:
    post:
      summary: Reject a subordinate settlement request
      description: Moves a manageable REQUESTED settlement request to REJECTED and marks linked commission entries REJECTED.
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/SettlementRequestIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettlementActionRequest'
      responses:
        '200':
          description: Settlement request rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementActionResponse'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/settlement-requests/{settlementRequestId}/mark-paid:
    post:
      summary: Mark an approved subordinate settlement request as paid
      description: Moves a manageable APPROVED settlement request to PAID and marks linked commission entries PAID.
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/SettlementRequestIdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettlementActionRequest'
      responses:
        '200':
          description: Settlement request marked paid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementActionResponse'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/hq-notices:
    get:
      summary: HQ notices visible to the leader
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/LeaderPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/notices/send-summary:
    get:
      summary: Notice send page recipient summary
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/LeaderPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/notices:
    get:
      summary: Notice send history page data
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/LeaderPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/notices/history:
    get:
      summary: Alias for notice send history page data
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/LeaderPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/notices/send:
    post:
      summary: Send or schedule a leader notice to subordinate partners and merchants
      description: Stores the notice body, recipient delivery rows, and activity log in one transaction. scheduledAt creates SCHEDULED/PENDING history; otherwise the notice is stored as SENT/SENT.
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoticeSendRequest'
      responses:
        '200':
          description: Notice send request was stored.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NoticeSendResponse'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/profile:
    get:
      summary: Leader profile page data
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/LeaderPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/activity-logs:
    get:
      summary: Leader activity log page data
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/CountryScopeQuery'
      responses:
        '200':
          $ref: '#/components/responses/LeaderPageData'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/signup-applications/{applicationId}/approve:
    post:
      summary: Approve a partner or merchant signup application
      description: Creates or connects runtime users plus the approved partners/merchants, distributor_contracts, and distributor_wallet_addresses records for a decisionable signup application.
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/ApplicationIdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupApprovalDecisionRequest'
      responses:
        '200':
          description: Signup application approved and runtime authority activated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupApprovalResponse'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
  /api/leader/signup-applications/{applicationId}/reject:
    post:
      summary: Reject a partner or merchant signup application
      description: Moves a decisionable signup application to REJECTED and records the decision activity.
      parameters:
        - $ref: '#/components/parameters/LeaderIdHeader'
        - $ref: '#/components/parameters/CountryScopesHeader'
        - $ref: '#/components/parameters/ApplicationIdPath'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupApprovalDecisionRequest'
      responses:
        '200':
          description: Signup application rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupRejectionResponse'
        '403':
          $ref: '#/components/responses/ForbiddenCountryScope'
...
components:
  parameters:
    LeaderIdHeader:
      name: X-Leader-Id
      in: header
      required: true
      schema:
        type: integer
        format: int64
      description: Temporary leader partner id until JWT auth is wired.
...
    CountryScopesHeader:
      name: X-Country-Scopes
      in: header
      required: true
      schema:
        type: string
      description: Temporary comma-separated country scopes until JWT auth is wired.
    CountryScopeQuery:
      name: countryScope
      in: query
      required: true
      schema:
        type: string
        minLength: 2
        maxLength: 2
      example: KR
    PeriodQuery:
      name: period
      in: query
      required: false
      schema:
        type: string
        pattern: '^\d{4}-\d{2}$'
      example: '2026-06'
    ApplicationIdPath:
      name: applicationId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    SettlementRequestIdPath:
      name: settlementRequestId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
  responses:
    ForbiddenCountryScope:
      description: Requested countryScope is outside the authenticated role's server-side country scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    LeaderPageData:
      description: Page data shaped for the current KORION PAY leader admin frontend.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/LeaderPageDataResponse'
...
  schemas:
...
    SignupApprovalDecisionRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 2000
        requestId:
          type: string
          maxLength: 128
...
    SettlementCreateRequest:
      type: object
      required:
        - periodStart
        - periodEnd
        - requestedAmount
      properties:
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        requestedAmount:
          type: number
          minimum: 0.000000000000000001
        walletAddressId:
          type: integer
          format: int64
          nullable: true
        memo:
          type: string
          maxLength: 2000
        requestId:
          type: string
          maxLength: 128
    SettlementActionRequest:
      type: object
      properties:
        amount:
          type: number
          minimum: 0.000000000000000001
          nullable: true
        reason:
          type: string
          maxLength: 2000
        requestId:
          type: string
          maxLength: 128
    SettlementActionResponse:
      type: object
      properties:
        settlementRequestId:
          type: integer
          format: int64
        requestNo:
          type: string
        recipientType:
          type: string
          enum:
            - LEADER
            - PARTNER
            - MERCHANT
        recipientPartnerId:
          type: integer
          format: int64
          nullable: true
        recipientMerchantId:
          type: integer
          format: int64
          nullable: true
        requestedAmount:
          type: number
        approvedAmount:
          type: number
        heldAmount:
          type: number
        status:
          type: string
          enum:
            - REQUESTED
            - APPROVED
            - REJECTED
            - PAID
        resultCode:
          type: string
          enum:
            - SETTLEMENT_REQUEST_CREATED
            - SETTLEMENT_REQUEST_APPROVED
            - SETTLEMENT_REQUEST_REJECTED
            - SETTLEMENT_REQUEST_PAID
        messageKey:
          type: string
...
    NoticeSendRequest:
      type: object
      required:
        - title
        - noticeType
        - body
        - channel
        - recipients
      properties:
        title:
          type: string
          maxLength: 255
        noticeType:
          type: string
          maxLength: 64
        body:
          type: string
        channel:
          type: string
          enum:
            - IN_APP
            - EMAIL
            - SMS
            - PUSH
        scheduledAt:
          type: string
          format: date-time
          nullable: true
          description: Must be later than the server clock when provided.
        recipients:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/NoticeRecipientRequest'
        requestId:
          type: string
          maxLength: 128
    NoticeSendResponse:
      type: object
      properties:
        noticeId:
          type: integer
          format: int64
        status:
          type: string
          enum:
            - SENT
            - SCHEDULED
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        sentAt:
          type: string
          format: date-time
          nullable: true
        recipientCount:
          type: integer
        resultCode:
          type: string
          enum:
            - NOTICE_SENT
            - NOTICE_SEND_SCHEDULED
        messageKey:
          type: string
...
    LeaderDashboardResponse:
      type: object
      required:
        - leaderProfile
        - kpis
        - organizationSummary
        - monthlyVolume
        - feeSummary
        - riskAlerts
      properties:
        leaderProfile:
          $ref: '#/components/schemas/LeaderProfile'
        kpis:
          $ref: '#/components/schemas/LeaderKpis'
        organizationSummary:
          $ref: '#/components/schemas/OrganizationSummary'
        monthlyVolume:
          type: array
          items:
            $ref: '#/components/schemas/MonthlyVolume'
        feeSummary:
          $ref: '#/components/schemas/FeeSummary'
        riskAlerts:
          type: array
          items:
            $ref: '#/components/schemas/RiskAlert'
    LeaderProfile:
      type: object
      properties:
        leaderId:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        loginId:
          type: string
        status:
          type: string
          enum:
            - COUNTRY_LEADER_APPROVED
        countryScopes:
          type: array
          items:
            type: string
...
    LeaderPartnerResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PartnerSummary'
        page:
          $ref: '#/components/schemas/PageMeta'
...